backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.4.1
chardet==5.2.0
apscheduler==3.10.4
httpx==0.27.0
pydantic==2.6.4
pydantic-settings==2.2.1
